Support functions for sorted menu feature
In generic shells the wildcard expansion produces an alphabetically ASC
sorted list. In grubscript there is no sorting, the items are generated in
so called "directory order" ('ls -U'), as they appear in the fs. Here we
introduce sorting based on file mtime DESC first, literal filename ASC
next.

In a typical scenario when a user puts a newly released iso into some dir
the new iso would naturally pop to the top in the dir's menu. In case there
is some iso in multiple flavours for the same version, a user should touch
those with the same mtime for the group to be sorted alphabetically.

When storing isos not sequentially it is a good idea to touch them with
their release date or mtime of some file from inside the iso like the
kernel or initrd.

Points of interest in grub sources:
- grub-core/commands/wildcard.c: wildcard_expand -> match_files -> fs->fs_dir
- grub-core/fs/fat.c: grub_fat_dir -> grub_fat_iterate_dir_next