Protect $root when passing control with 'configfile'
Some of inc-*.cfg files load foreign configs with 'configfile'. Activating
such menu item and then pressing 'Esc' we would "browse" back to the "page"
for that distro. If we set $root to the loop for 'configfile' but not
restore it later, all menu items on the "page" become useless.